backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
supabase>=2.3.0
python-jose>=3.3.0
web3>=6.15.0
pytest>=8.0.0
httpx>=0.26.0
playwright>=1.41.0
razorpay>=1.3.0
python-multipart>=0.0.9
email-validator>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
